#moved register extensions to proper file
